# Food-Delivery-Microservice/Node-Microservices/ReviewService/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Order API",
    "description": "API for managing Orders",
    "version": "1.0.0"
  },
  "host": "localhost:5300/api/review",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "paths": {
    "/create": {
      "post": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/restaurantId/{restaurantId}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "restaurantId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/delete/{reviewId}": {
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "reviewId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  }
}